refactor: cleaned up folder structure (#1228)
