+++ src/app/core/pwa-install/pwa-install.component.scss
