+++ src/app/core/sync-status/sync-status/sync-status.component.spec.ts
@@ -30,7 +30,7 @@ import {
   EntityRegistry,
   entityRegistry,
 } from "../../entity/database-entity.decorator";
-import { expectObservable } from "../../../test-utils/observable-utils";
+import { expectObservable } from "../../../utils/test-utils/observable-utils";
 
 describe("SyncStatusComponent", () => {
   let component: SyncStatusComponent;
