+++ src/app/core/ui/ui.module.ts
@@ -41,7 +41,7 @@ import { PermissionsModule } from "../permissions/permissions.module";
 import { EntityUtilsModule } from "../entity-components/entity-utils/entity-utils.module";
 import { TranslationModule } from "../translation/translation.module";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
-import { PwaInstallModule } from "app/pwa-install/pwa-install.module";
+import { PwaInstallModule } from "app/core/pwa-install/pwa-install.module";
 
 /**
  * The core user interface structure that ties different components together into the overall app layout.
